institutional_trap_v3/requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
orjson>=3.9.0
aiologger>=0.7.0
joblib>=1.3.0
aiohttp>=3.8.0